WT-7974: Fix JS fitToWindow() ignoring parent.
The fitToWindow() function was making the assumption that the given
widget would be visible inside the whole window. This is, however, not
always true for widgets that are not a direct child of the domroot. This
lead to the widget being cut off at the edge of its parent widget. This
was especially visible for popups using WWidget::positionAt(), which
are reparented to the first parent, of the widget it position against,
that has visible scrollbar.

The fitToWindow() function will now try to position the widget inside
the intersection between its parent viewport and the window's viewport.